discord get_schedule
