+++ Backend/discord/views.py
@@ -41,7 +41,7 @@ def connect_study(request, study_id):
 
 @api_view(['GET'])
 def study_schedule_list(request, study_id):
-    
+    print("==================")
     guild_id = request.query_params.get("guild_id")
     mapping = get_object_or_404(DiscordStudyMapping, guild_id=guild_id, study_id=study_id)
     study = Study.objects.get(id=study_id)
