+++ bot/practice.py
@@ -75,64 +75,5 @@ async def connect(ctx, study_id: int):
     if data:
         print(data)
 
-@bot.command(name="study_schedule_list")
-async def study_schedule_list(ctx, study_id):
-    guild = ctx.guild
-    url = f"{DJANGO_API_URL}studies/{study_id}/discord/study_schedule_list/"
-    params = {
-        "guild_id": guild.id,
-        "guild_name": guild.name,
-    }
-
-    async with aiohttp.ClientSession() as session:
-        async with session.get(url, params=params) as resp:
-            if resp.status == 200:
-                study = await resp.json()
-            else:
-                study = None
-    if study:
-        name = study["name"]
-        message = f"## 스터디: { name }\n"
-        for schedule in study["schedule"]:
-            message += (f"### 제목: {schedule.get('schedule').get('title')}\n"
-                        f"- 작성자: {schedule.get('author').get('username')}\n"
-                        f"\t- 내용: {schedule.get('schedule').get('description')}\n"
-                        f"\t- 시작: {' '.join(schedule.get('schedule').get('start_at')[:-4].split('T'))}\n"
-                        f"\t- 종료: {' '.join(schedule.get('schedule').get('end_at')[:-4].split('T'))}\n"
-                        )
-        await ctx.send(message)
-    else:
-        ctx.send("스터디 일정이 존재하지 않습니다.")
-
-
-@bot.command(name="guild_schedule_list")
-async def guild_schedule_list(ctx):
-    guild = ctx.guild
-    url = f"{DJANGO_API_URL}discord/guild_schedule_list/"
-    params = {
-        "guild_id": guild.id,
-        "guild_name": guild.name,
-    }
-    async with aiohttp.ClientSession() as session:
-        async with session.get(url, params=params) as resp:
-            if resp.status == 200:
-                data = await resp.json()
-            else:
-                data = None
-    if data:
-        for study in data:
-            name = study["name"]
-            message = f"## 스터디: { name }\n"
-            for schedule in study["schedule"]:
-                message += (f"### 제목: {schedule.get('schedule').get('title')}\n"
-                            f"- 작성자: {schedule.get('author').get('username')}\n"
-                            f"\t- 내용: {schedule.get('schedule').get('description')}\n"
-                            f"\t- 시작: {' '.join(schedule.get('schedule').get('start_at')[:-4].split('T'))}\n"
-                            f"\t- 종료: {' '.join(schedule.get('schedule').get('end_at')[:-4].split('T'))}\n"
-                            )
-            await ctx.send(message)
-    else:
-        ctx.send("스터디 일정이 존재하지 않습니다.")
-
 if __name__ == "__main__":
     bot.run(TOKEN)
